+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 0.1.0 (beta.1)
 
 - Introduced your friendly bank, The Lethal Credit Union
-- Add deposit, withdraw and balance functionality
+- balance, deposit, withdraw and credit banking functionality
 - Deposit scrap on the ship into the bank (less loot concept)
-- Withdraw from the bank into a dollar stack
-- Check your balance
+- Withdraw from the bank into some LCU buck
+- Auto-bank scrap at end of game round